rearchitect the encoder and encode scheme
