+++ src/command.rs
@@ -2,6 +2,7 @@ pub mod args;
 pub mod auto_encode;
 pub mod crf_search;
 pub mod encode;
+pub mod encoders;
 pub mod print_completions;
 pub mod sample_encode;
 pub mod vmaf;
